requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
dconfig==1.2.0
Django==2.2.17
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.1
idna==2.10
PyJWT==2.1.0
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
